factor out JvmCompilationTask extensions to reuse in abi generation
